Remove potential circular import path
The Mixin classes were in io, and imported within the vrt module,
but the io module also imported the vrt module (within _io). This
worked when copy was being imported before io in __init__, but removing
that import from __init__ produced a circular import.
